requirements.txt
absl-py==2.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
audioread==3.0.1
av==14.2.0
babel==2.17.0
billiard==4.2.1
blis==0.7.11
catalogue==2.0.10
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.0
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.1
coqpit-config==0.2.0
coqui-tts==0.26.0
coqui-tts-trainer==0.2.3
ctranslate2==4.5.0
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dateparser==1.1.8
decorator==5.2.1
Django==5.2.1
docopt==0.6.2
einops==0.8.1
elevenlabs==1.55.0
encodec==0.1.1
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
grpcio==1.71.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
inflect==7.5.0
Jinja2==3.1.6
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.8
kombu==5.5.3
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
monotonic-alignment-search==0.1.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.3.0
murmurhash==1.0.12
networkx==3.4.2
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.21.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
preshed==3.0.9
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
pyparsing==3.2.3
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==76.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
SudachiDict-core==20250129
SudachiPy==0.6.10
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.50.3
triton==3.2.0
typeguard==4.4.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
